Add files via upload
In this update each article is assigned a number from 1 to 10. When user launches the program, they are prompted to enter the article number they wish to read, ranging from 1 to 10. Inputing the desired article number will instantly open in their browser.
If the user decide they don't want to read any of the articles, they can simply input '0,' and the program will gracefully exit.